requirements.txt
hydra-core
matplotlib
seaborn
transformers
wandb
datasets
omegaconf
tqdm
einops